+++ Tests/Services/GetEnvironmentVariables/GetEnvironmentVariablesValidate.cs
@@ -4,7 +4,7 @@ using Xunit;
 using App.Entities;
 using App.Services;
 
-namespace Tests 
+namespace Tests.Services
 {
     public class GetEnvironmentVariablesValidate
     {
