+++ Tests/Services/OperationCheckerServiceIsOperation.cs
@@ -1,7 +1,7 @@
 using App.Services;
 using Xunit;
 
-namespace Tests
+namespace Tests.Services
 {
     public class OperationCheckerServiceIsOperation
     {
