+++ Tests/Services/Validations/Level3/HttpVerbsValidate.cs
@@ -4,7 +4,7 @@ using App.Services.Validations.Level3;
 using App.Entities;
 using App.Entities.Swagger.Two;
 
-namespace Tests
+namespace Tests.Services.Validations.Level3
 {
     public class HttpVerbsValidate
     {
